Moved api rest logic to rest folder
